Add tests and refactor trollstalker a bit
